requirements.txt
ultralytics
torch
pyserial
opencv-python
simple-pid